Add files via upload
Data collected by healthcare organizations will be highly dimensional since each patient will have their own history of medication. Therefore, prediction is a challenging task when dealing with such high-dimensional data. The purpose of this project is to create  classifiers that could model the medical data. The study uses two sets of medical data: one set of data is related to Heart Disease and the other is Eye Infection Data. For prediction, we have used some of the machine learning methods such as Naive Bayesian, Decision Tree, Multilayer Perceptron, Random Forest, Support Vector Machine and , Decision Tree,XGBoost and AdaBoost  along with some Ensemble methods like Bagging and Voting that helps in increasing the model’s performance. For validating the classifier, both split and cross-validation techniques are used, and well-known evaluation measures for predictive models like accuracy, recall, precision,ROC and F-measure are considered. In this work, we analyzed the data in two stages. In the first stage, we make use of all the available features while in the second we select the features which have a positive correlation. Along with the traditional machine learning algorithms, the adopted approach for deep learning, particularly Convolutional Neural Network (CNN), is used to monitor the results. Better predictions are made in the second stage, according to the experimental findings. Additionally, research shows that split percentage creates better predictive models. 

As a novelty, we have preprocessed the given datasets using z-score normalization to remove outliers. The technique GridSearchCV is implemented to find the best parameters that give better predictions for all the classifiers. In addition XGBoost, AdaBoost and ANN models are implemented to classify the data. Using 80- 20 splits in the second stage, XGBoost achieved the highest accuracy for the Eye Infection dataset and Heart Disease dataset.While in the first stage XGBoost achieved highest accuracy for Heart Disease Dataset and AdaBoost topped all the other algorithms for Eye Infection Dataset.  Due to the quantity of the analyzed datasets, the CNN and ANN  model performed badly.